Update to the draft paper
- Split each section into a separate file

- Rewrite the description of the general framework: There is no major
change to the framework itself, except for allowing primitive types to
have more than one index (so that we can have numbers indexed by a
scaling factor and an offset, for instance). I have rewritten most of
the explanation though, in the hope that it is more understandable
now. A major change to the presentation is that I have changed the
order of syntax and semantics to be interleaved: I define the
semantics of types (index-erasure and relational) before the syntax of
programs. This is motivated by wanting to present the idea of
considering the meanings of types as determining the syntax of
programs, rather than the other way round.

- Put (a plan for) more stuff into the motivating examples section.